Assignment For Module 9
Update models, routes and index.js for all three entities. The two for the project that I would like you to consider for grading are the User entity and the Post entity. Still trying to figure out how to test routes, will resubmit with corrections if needed. Only part I am really worried about is lines 21-23 in index.js for app.use.